Added data files to rebuild the KANJIS table
Also added a schema with table definitions as they were used in the "Trinity" alpha version. These may be helpful to build simple vocabulary or kanji reading reviews, and many other features that require looking up vocabulary and kanji data.

The Trinity table definitions are added for reference only. The scripts require more work to be added into the RevTK repository, I'll provide an archive of the data files that can be loaded in the database, short of fixing the scripts.