Moved the maintenance page within the /web folder (site-specific). Updated .htaccess with 302 HTTP Status Code.
I wanted to use HTTP 503 (Service Temporarily Unavailable), but for some reason the web host returns an Apache template, instead of using /maintenance.html, so I reverted to HTTP 302. 302 isn't great because the url changes, if the user reloads the page to check that the website is back up, it doesn't work because they reload the maintenance page instead.