Fix Study search with dots in keyword, cleaned up Study action a little.
There were two problems with dots in the keyword search. One problem was
due to the default sfPatternRouting using dots as separator, this is
fixed by adding a "requirements" to the route to match the dots as part
of the parameter. This fixes the route going to the default 404 action.

The second problem is less clear. The web host seems to filter url
patterns which I assume look like exploits on known packages such as
PunBB (one page was blocked there before). The 404 page FROM THE WEB
HOST would show there, with keywords using multiple dots such as "made
in..." or other uncommon combinations eg. "Bldg. [old]". The fix was to
replace all of these with dashes and use wildcards on the backend
(similar to handling pretty url headlines).